+++ test/correctness/PySys_internal_125/run.py
@@ -24,7 +24,7 @@ class PySysTest(BaseTest):
 		reasonFailedEvalAssert = self.getOutcomeReason()
 		self.addOutcome(PASSED, override=True)
 
-		self.assertThat('actual == expected', actual=r"prefix f\oo bar suffix", expected=r'f\oobar')
+		self.assertThat('actual == expected', actual=r"prefix f\oo bar suffix", expected=r'f\oobar', failureOutcome=BADPERF)
 		reasonFailedAssert = self.getOutcomeReason()
 		self.addOutcome(PASSED, override=True)